# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de Reservas de Club Deportivo
  description: |
    API REST para gestionar canchas, socios y reservas del Club Deportivo Encuentro.

    El contrato sigue el enunciado del trabajo práctico. La URL base puede variar
    según el despliegue (por ejemplo, `/api` o `/v1`).
  version: 1.0.0
servers:
  - url: /
    description: Servidor local de desarrollo
...
paths:
...
  /canchas:
    get:
      summary: Listar canchas
      description: |
        Devuelve el listado paginado de canchas. Se pueden aplicar filtros por
        deporte, nombre (búsqueda parcial e insensible a mayúsculas), techada y
        activa. Sin filtros se devuelven canchas activas e inactivas.
      tags:
        - Canchas
      parameters:
        - $ref: '#/components/parameters/IdDeporteFilter'
        - $ref: '#/components/parameters/NombreFilter'
        - $ref: '#/components/parameters/TechadaFilter'
        - $ref: '#/components/parameters/ActivaFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/GetCanchasList200'
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Crear una cancha
      tags:
        - Canchas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanchaCreate'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /canchas/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Obtener una cancha por ID
      tags:
        - Canchas
      responses:
        '200':
          $ref: '#/components/responses/GetCancha200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Actualizar parcialmente una cancha
      tags:
        - Canchas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanchaUpdate'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Eliminar una cancha
      description: |
        Elimina una cancha solo si no tiene reservas asociadas. Si tiene
        reservas, responde 409 y podrá desactivarse mediante PATCH.
      tags:
        - Canchas
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /canchas/disponibles:
    get:
      summary: Consultar canchas disponibles en un intervalo
      description: |
        Devuelve las canchas activas que están libres durante todo el intervalo
        solicitado. No crea ni retiene la reserva.
      tags:
        - Canchas
      parameters:
        - $ref: '#/components/parameters/FechaRequired'
        - $ref: '#/components/parameters/HoraInicioRequired'
        - $ref: '#/components/parameters/HoraFinRequired'
        - $ref: '#/components/parameters/IdDeporteFilter'
        - $ref: '#/components/parameters/TechadaFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/GetCanchasList200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Cancha:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único de la cancha
          example: 1
        nombre:
          type: string
          description: Nombre o denominación de la cancha
          example: "Cancha 1 - Fútbol 5"
        id_deporte:
          type: integer
          description: Identificador del deporte asociado
          example: 1
        precio_hora:
          type: integer
          description: Tarifa vigente por hora en centavos
          example: 1000000
        techada:
          type: boolean
          description: Indica si la cancha es cubierta
          example: false
        activa:
          type: boolean
          description: Indica si admite nuevas reservas
          example: true
    CanchaCreate:
      type: object
      required:
        - nombre
        - id_deporte
        - precio_hora
      properties:
        nombre:
          type: string
          description: Nombre de la cancha
          example: "Cancha 1 - Fútbol 5"
        id_deporte:
          type: integer
          description: Identificador del deporte asociado
          example: 1
        precio_hora:
          type: integer
          description: Tarifa por hora en centavos. Debe ser mayor a cero.
          minimum: 1
          example: 1000000
        techada:
          type: boolean
          description: Indica si la cancha es cubierta
          default: false
          example: false
        activa:
          type: boolean
          description: Indica si la cancha admite nuevas reservas
          default: true
          example: true
    CanchaUpdate:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la cancha
          example: "Cancha 1 - Fútbol 5"
        precio_hora:
          type: integer
          description: Tarifa por hora en centavos. Debe ser mayor a cero.
          minimum: 1
          example: 1200000
        techada:
          type: boolean
          description: Indica si la cancha es cubierta
          example: true
        activa:
          type: boolean
          description: Indica si la cancha admite nuevas reservas
          example: true
...
    Links:
      type: object
      description: Enlaces HATEOAS para navegar entre páginas
      properties:
        _first:
          $ref: '#/components/schemas/Link'
        _prev:
          $ref: '#/components/schemas/Link'
        _next:
          $ref: '#/components/schemas/Link'
        _last:
          $ref: '#/components/schemas/Link'
    Link:
      type: object
      properties:
        href:
          type: string
          description: URI para navegar a la página indicada
          example: "http://localhost:5000/canchas?_offset=0&_limit=10"
    Error:
      type: object
      properties:
        code:
          type: string
          description: Código alfanumérico del error
          example: "ERROR_VALIDACION"
        message:
          type: string
          description: Resumen breve del error
          example: "El cuerpo de la solicitud es inválido"
        level:
          type: string
          description: Nivel del error
          enum:
            - info
            - warning
            - error
          example: "error"
        description:
          type: string
          description: Descripción detallada del error
          example: "El campo 'precio_hora' debe ser un entero mayor a cero"
    Errores:
      type: object
      properties:
        errors:
          type: array
          description: Listado de errores
          items:
            $ref: '#/components/schemas/Error'
...
    CanchasListResponse:
      type: object
      properties:
        canchas:
          type: array
          items:
            $ref: '#/components/schemas/Cancha'
        _links:
          $ref: '#/components/schemas/Links'
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Identificador entero positivo del recurso. Si el valor no es válido, el servicio responde 400.
      schema:
        type: integer
        minimum: 1
      example: 1
    Limit:
      name: _limit
      in: query
      description: Cantidad máxima de registros por página. Debe ser un entero entre 1 y 100; de lo contrario se responde 400.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
    Offset:
      name: _offset
      in: query
      description: Desplazamiento desde el primer registro. Debe ser un entero mayor o igual a 0; si no, se responde 400.
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
    NombreFilter:
      name: nombre
      in: query
      description: Búsqueda parcial e insensible a mayúsculas por nombre
      schema:
        type: string
      example: "Cancha 1"
    IdDeporteFilter:
      name: id_deporte
      in: query
      description: Filtrar por identificador de deporte
      schema:
        type: integer
      example: 1
    TechadaFilter:
      name: techada
      in: query
      description: Filtrar por indicador de cancha techada
      schema:
        type: boolean
      example: true
    ActivaFilter:
      name: activa
      in: query
      description: Filtrar por indicador de cancha activa
      schema:
        type: boolean
      example: true
...
    FechaRequired:
      name: fecha
      in: query
      required: true
      description: Día para consultar disponibilidad
      schema:
        type: string
        format: date
      example: "2026-10-15"
    HoraInicioRequired:
      name: hora_inicio
      in: query
      required: true
      description: Hora de inicio del intervalo
      schema:
        type: string
        pattern: '^([01]\d|2[0-3]):00:00$'
      example: "18:00:00"
    HoraFinRequired:
      name: hora_fin
      in: query
      required: true
      description: Hora de fin del intervalo
      schema:
        type: string
        pattern: '^([01]\d|2[0-3]):00:00$'
      example: "20:00:00"
...
  responses:
...
    GetCanchasList200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CanchasListResponse'
    GetCancha200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Cancha'
...
    Created:
      description: Creado
...
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'
    Conflict:
      description: Conflicto de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'
    NoContent:
      description: Sin contenido
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'